fix: move lazydict out to its own file.
