+++ python/pyiceberg/avro/decoder_fast.pyx
@@ -19,7 +19,7 @@ import cython
 from cython.cimports.cpython import array
 from datetime import datetime, time
 from uuid import UUID
-from pyiceberg.avro.reader import LazyDictIntInt
+from pyiceberg.avro.lazydict import LazyDictIntInt
 from pyiceberg.avro import STRUCT_DOUBLE, STRUCT_FLOAT
 from pyiceberg.utils.datetime import micros_to_time, micros_to_timestamp, micros_to_timestamptz
 from pyiceberg.utils.decimal import unscaled_to_decimal
